requirements.txt
flask==2.3.3
gunicorn==21.2.0
werkzeug==2.3.7
pymongo==4.6.1
python-dotenv==1.0.0
dnspython==2.4.2 